requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
Django==3.2.20
django-admin-interface==0.26.0
django-admin-list-filter-dropdown==1.0.3
django-adminlte3==0.1.6
django-allauth==0.63.0
django-autoslug==1.9.9
django-colorfield==0.9.0
django-crispy-forms==2.0
django-rest-framework==0.1.0
django-social-share==2.3.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
oauthlib==3.2.2
openpyxl==3.1.2
Pillow==9.4.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==1.26.6
